requirements.txt
sentence-transformers  ## opensource embedding model
langchain
flask
pypdf
python-dotenv
pinecone[grpc]       
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .